+++ History.md
@@ -1,4 +1,5 @@
 ## V 4.4.8
+- animation new method "setSubframe" to enable subframe rendering (true by default).
 - hidden guided layers and parenting restore
 - split animations export resetting segments